requirements-dev.txt
Sphinx
sphinx-rtd-theme
django-debug-toolbar
behave-django
selenium